Add authentication to updater
This adds some authentication to the updater using a shared secret in `config.php`. The whole process is cookieless and thus if somebody closes the tab all authentication is gone (which is desired from a security PoV)

To test this define an `updater.secret` in `config.php` and try to access it with a valid or invalid token.

The next step is to add the automagic login from core once this is in. We can do that using the new updatenotification app at https://github.com/owncloud/core/pull/22238.