Fix setting of maintenance mode from within PHP in CLI mode
* fixes #210

Background:

The `setMaintenanceMode` does `require('config.php');` and thus is loading the config.php into the CLI context, changing the maintenance mode value and then writing the file back. This works fine in the web based UI because there each step (disabling the maintenance mode is a single step) is executed in one request. On the CLI side we do all steps within one single PHP call. That results in following execution:

* fetching new code
* replacing code
* enable maintenance mode
* upgrade
* disable maintenance mode

This sounds not like a problem, but there is one in the disable step. The "upgrade" step updates the config.php outside of the currently running PHP process, but the "disable maintenance mode" just uses the same config.php as when the CLI process was started. That means that it still had the old version number and thus was writing it back after the upgrade happened. This triggers the described misbehavior in #210.

This is only triggered sometimes because we use OPCache to cache PHP files, which is checked once a second for updates, but if between the end of "upgrade" and beginning of "disable maintenance mode" this is not checked it loads the old file content. That's the reason why also some tests fail from time to time in a weird way.

Signed-off-by: Morris Jobke <hey@morrisjobke.de>