Add a configuration option to skip the integrity verification
On systems with very low PHP memory (e.g. 64MB), the PHP function
openssl_verify may fail as it checks the complete nextcloud zip file
in RAM which is not possible with only 64 MB.
For this, an option has been introduced that allows to skip the
integrity verification.
Note: This option should only be used for systems that have issues
in running the updater.

Signed-off-by: Mathias K <koehrer08@arcor.de>