Generate URLs via JavaScript
URL generation via PHP do not consider cases where the application URL at a reverse proxy is different from the local URL.

Example:
- Public: home.acme/nextcloud
- Private: 192.168.1.8

A reverse proxy forwards the request internally. As the updater is not installed in a subdirectory, the generated (via PHP) URLs are wrong. nextcloud/server has all the code to generate proper URLs for such cases. However, this code is not available here.

Generating the URLs via JavaScript is simpler as the browser is already visiting the right page.

Signed-off-by: Daniel Kesselberg <mail@danielkesselberg.de>